bug: load a page of results only. Not the full result set

+++ src/main/java/de/trustable/ca3s/core/service/dto/UserDTO.java
@@ -60,7 +60,7 @@ public class UserDTO {
 
     private Long tenantId;
 
-    private Integer failedLogins;
+    private Long failedLogins;
 
     private Instant blockedUntilDate = null;
 
@@ -235,11 +235,11 @@ public class UserDTO {
         this.tenantId = tenantId;
     }
 
-    public Integer getFailedLogins() {
+    public Long getFailedLogins() {
         return failedLogins;
     }
 
-    public void setFailedLogins(Integer failedLogins) {
+    public void setFailedLogins(Long failedLogins) {
         this.failedLogins = failedLogins;
     }
